AgenticAi/pyproject.toml
[project]
name = "agenticai"
version = "0.1.0"
description = "I am learning Agentic Ai"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-tavily>=0.2.8",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.3.4",
    "langsmith>=0.4.5",
    "python-dotenv>=1.1.1",
]
